Adding juliatestloop.jl for bl data.
First prototype of asynchronous solving. Includes scripts for first and second stage solves, shared info to parse to those models, and data from a test run of the loop file. In the cluster, this file will be re-constructed with individual job submissions in mind. Next step is to check/adjust for wsgep.